Separated containers from screens
