+++ src/package.json
@@ -37,6 +37,7 @@
   },
   "scripts": {
     "test": "node_modules/.bin/chimp --watch --compiler=coffee:coffee-script/register --ddp=http://localhost:3000 --path=tests/cucumber/features",
+    "testOnce": "node_modules/.bin/chimp --compiler=coffee:coffee-script/register --ddp=http://localhost:3000 --path=tests/cucumber/features",
     "build": "meteor build --server=http://stg.arciau.lt ../build",
     "testService": "meteor test-packages packages/carpool-service --settings settings.json",
     "lint": "eslint ."
