+++ src/packages/carpool-view-flow/react/containers/RidesListWithData.jsx
@@ -1,5 +1,5 @@
 import { createContainer } from 'meteor/react-meteor-data';
-import RidesList from '../../components/rides-list/RidesList';
+import RidesList from '../components/rides-list/RidesList';
 
 /*global Progress*/
 /*global carpoolService*/
@@ -13,21 +13,14 @@ export default createContainer(({filterOwn = "all", role = "driver", aLoc, bLoc,
     bTime: bTime && bTime.toDate()
   }
   let trips
-  //d("RideOffersList query", bTime);
+
   // Some magic here to remove undefined values
   Object.keys(query).forEach((key)=>{ query[key] || delete query[key] });
 
-  //console.log("Filter query:", query, aLoc)
   if("your" == filterOwn) {
     trips = carpoolService.pullOwnTrips(query, progress.setProgress.bind(progress, 'ownTrips'));
-    if(100 == progress.getProgress()) {
-       //d(`Own ${role} trips:`, trips);
-    }
   } else {
     trips = carpoolService.pullActiveTrips(query, progress.setProgress.bind(progress, 'activeTrips'));
-    if(100 == progress.getProgress()) {
-      //d(`All ${role} trips:`, trips);
-    }
   }
 
   return {
