+++ src/packages/carpool-view-flow/react/screens/RideOffers.jsx
@@ -1,18 +1,11 @@
 import React from 'react';
-import { createContainer } from 'meteor/react-meteor-data';
-import RidesList from '../../components/rides-list/RidesList';
-import { connect } from 'react-redux';
-import Loader from '../../components/common/Loader.jsx';
-import locationFromSelector from '../../redux/selectors/locationFrom.js';
-import locationToSelector from '../../redux/selectors/locationTo.js';
-import tripDateTimeSelector from '../../redux/selectors/tripDateTime.js';
+import RidesList from '../components/rides-list/RidesList';
+import Loader from '../components/common/Loader.jsx';
 import { FlatButton, Snackbar, Divider } from 'material-ui';
 import { StyleSheet, css } from 'aphrodite';
 
-import { createDrive, createRide } from '../../redux/modules/general/actions.js';
+import { createDrive, createRide } from '../redux/modules/general/actions.js';
 
-/*global Progress*/
-/*global carpoolService*/
 /*global FlowRouter*/
 
 const styles = StyleSheet.create({
@@ -33,7 +26,7 @@ const styles = StyleSheet.create({
   },
 })
 
-class RideOffersScreen extends React.Component {
+export default class RideOffersScreen extends React.Component {
   constructor (props) {
     super(props);
 
@@ -114,42 +107,5 @@ RideOffersScreen.propTypes = {
   locationFrom: React.PropTypes.string.isRequired,
   locationTo: React.PropTypes.string.isRequired,
   tripDateTime: React.PropTypes.object.isRequired,
+  dispatch: React.PropTypes.func.isRequired,
 }
-
-const connectedRideOffersScreen = connect((state) => ({
-  locationFrom: locationFromSelector(state),
-  locationTo: locationToSelector(state),
-  tripDateTime: tripDateTimeSelector(state),
-}))(RideOffersScreen)
-
-export default createContainer(({filterOwn = "all", role = "driver", aLoc, bLoc, bTime}) => {
-  const progress = new Progress();
-  const query = {
-    role: role,
-    fromLoc: aLoc,
-    toLoc: bLoc,
-    bTime: bTime && bTime.toDate()
-  }
-  let trips
-  //d("RideOffersList query", bTime);
-  // Some magic here to remove undefined values
-  Object.keys(query).forEach((key)=>{ query[key] || delete query[key] });
-
-  //console.log("Filter query:", query, aLoc)
-  if("your" == filterOwn) {
-    trips = carpoolService.pullOwnTrips(query, progress.setProgress.bind(progress, 'ownTrips'));
-    if(100 == progress.getProgress()) {
-       //d(`Own ${role} trips:`, trips);
-    }
-  } else {
-    trips = carpoolService.pullActiveTrips(query, progress.setProgress.bind(progress, 'activeTrips'));
-    if(100 == progress.getProgress()) {
-      //d(`All ${role} trips:`, trips);
-    }
-  }
-
-  return {
-    progress,
-    trips
-  };
-}, connectedRideOffersScreen);
