+++ src/packages/carpool-view-flow/react/screens/YourDrive.jsx
@@ -1,19 +1,9 @@
 import React from 'react'
-import { createContainer } from 'meteor/react-meteor-data';
-
-import { _ } from 'meteor/underscore';
-
-import GoogleMap from '../../components/map/GoogleMap'
-import { config } from '../../config'
 import RaisedButton from 'material-ui/lib/raised-button';
-import TripInfoWithMap from '../../components/ride-info/TripInfoWithMap.jsx'
-import Loader from '../../components/common/Loader'
-import { getUserPicture } from '../../api/UserPicture.coffee'
 import { StyleSheet, css } from 'aphrodite'
-/*global Meteor*/
-/*global getUserName*/
-/*global carpoolService*/
-/*global Progress*/
+
+import TripInfoWithMap from '../components/ride-info/TripInfoWithMap.jsx'
+import Loader from '../components/common/Loader'
 
 const styles = StyleSheet.create({
   screenWrap: {
@@ -23,11 +13,9 @@ const styles = StyleSheet.create({
   },
 })
 
-class YourDrive extends React.Component {
+export default class YourDrive extends React.Component {
 
   render () {
-    const topBarHeight = 45
-    const mapHeight = 375
     const {progress, stops, itinerary, user, isRequested} = this.props;
 
     if (100 != progress.getProgress()) {
@@ -66,21 +54,3 @@ YourDrive.propTypes = {
   isRequested: React.PropTypes.object,
   itinerary: React.PropTypes.array,
 };
-
-export default createContainer(({tripId}) => {
-  const progress = new Progress();
-  const drive = carpoolService.pullOneTrip({_id: tripId}, progress.setProgress.bind(progress, 'oneTrip'));
-  const stops = carpoolService.pullStops(progress.setProgress.bind(progress, 'stops'));
-  const user = drive && Meteor.users.findOne({_id: drive.owner});
-
-  const itinerary = carpoolService.pullDriverItinerary(drive);
-  const isRequested = drive && _(drive.requests).findWhere({userId: Meteor.userId()});
-
-  return {
-    progress,
-    stops,
-    user,
-    itinerary,
-    isRequested
-  };
-}, YourDrive);
